+++ src/lib.rs
@@ -13,6 +13,7 @@
 #![warn(missing_docs)]
 #![warn(rustdoc::broken_intra_doc_links)]
 
+mod atexit;
 mod client;
 mod concurrency;
 mod container;